Experiment with int instead of uint8_t

+++ Source/decoder.cpp
@@ -143,14 +143,14 @@ namespace tga
 		}
 	}
 
-	uint8_t Decoder::read8()
+	//uint8_t Decoder::read8()
+	int Decoder::read8()
 	{
 		return m_file->read8();
 	}
 
 	// Read 16 bits using little-endian byte ordering.
 	uint16_t Decoder::read16()
-	//int Decoder::read16()
 	{
 		// TODO: List initialize.
 		uint8_t b1 = m_file->read8();
