+++ Source/encoder.hpp
@@ -25,7 +25,8 @@ namespace tga
 									const int height,
 									void (Encoder::*writePixel)(T));
 
-		void write8(uint8_t value);
+		//void write8(uint8_t value);
+		void write8(int value);
 		void write16(uint16_t value);
 		//void write16(int value);
 		void write24AsRgb(color c);
